add: hooks/useSacripts
